## Container
With styles in place, let's get our hands on extracting a shared layout for all future views to come.
Start by adding `container` parameter to `index` in `View`:

==> View.fs:26-31

and div with id "container" just after the div "header":

==> View.fs:33-38

`index` previously was a constant value, but it has now become a function taking `container` as parameter.

We can now define the actual container for the "home" page:

==> View.fs:10-12

For now it will only contain plain "Home" text.
Let's also extract a common function for creating the WebPart, parametrized with the container itself.
Add to `App` module, just before the `browse` WebPart the following:

==> App.fs:9-10

Usage for the home page looks like this:

==> App.fs:19-20

Next, containers for each valid route in our application can be defined in `View` module:

==> View.fs:10-24

Note that both `home` and `store` are constant values, while `browse` and `details` are parametrized with `genre` and `id` respectively.

`html` can be now reused for all 4 views:

==> App.fs:12-26