## Delete album
We can't make any operation on an album yet.
To fix this, let's first add the delete functionality:

==> View.fs:87-103

`deleteAlbum` is to be placed in the `View` module. It requires new markup functions:

==> View.fs:16-19

- `strong` is just an emphasis
- `form` is HTML element for a form with it's "method" attribute set to "POST"
- `submitInput` is button to submit a form

A couple of snippets to handle `deleteAlbum` are still needed, starting with `Db`:

==> Db.fs:41-46

for getting `Album option` (not `AlbumDetails`).
New route in `Path`:

==> Path.fs:16-18

Finally we can put following in the `App` module:

```fsharp
let deleteAlbum id =
    match Db.getAlbum id (Db.getContext()) with
    | Some album ->
        html (View.deleteAlbum album.Title)
    | None ->
        never
```

==> App.fs:65-65

Note that the code above allows us to navigate to to "/admin/delete/%d", but we still are not able to actually delete an album.
That's because there's no handler in our app to delete the album from database.
For the moment both GET and POST requests will do the same, which is return HTML page asking whether to delete the album.

In order to implement the deletion, add `deleteAlbum` to `Db` module:

==> Db.fs:48-50

The snippet takes an `Album` as a parameter - instance of this type comes from database, and we can invoke `Delete()` member on it - SQLProvider keeps track of such changes, and upon `ctx.SubmitUpdates()` executes necessary SQL commands. This is somewhat similar to the "Active Record" concept.

Now, in `App` module we can distinguish between GET and POST requests:

==> App.fs:43-55

- `deleteAlbum` WebPart gets passed the `choose` with two possibilities.
- `GET` and `POST` are WebParts that succeed (return `Some x`) only if the incoming HTTP request is of GET or POST verb respectively.
- after successful deletion of album, the `POST` case redirects us to the `Path.Admin.manage` page

> Important: We have to wrap both GET and POST handlers with a `warbler` - otherwise they would be evaluated just after `Some album` match, resulting in invoking `Db.deleteAlbum` even if POST does not apply.

The grid can now contain a column with link to delete the album in question:

==> View.fs:68-85

- there's a new empty column header in the first row
- in the last column of each album row comes a cell with link to delete the album
- note, we had to use the `yield` keyword again