# Introduction
This is a tutorial on how to create an application with [F#](http://fsharp.org) and [Suave.IO](http://suave.io) framework.
It's inspired by the Music Store tutorial created by the ASP.NET team [available here](http://www.asp.net/mvc/overview/older-versions/mvc-music-store/mvc-music-store-part-1).
Check out this link if you want to find out what the application is going to offer.

Target audience for this tutorial are mainly C# developers familiar with ASP.NET MVC, who want to learn how to write a real application in F#.
You can still benefit from the tutorial if you don't have C# / .NET background, however you may find some aspects not clear - From time to time there will be a comparison with how the same functionality could be written in ASP.NET MVC & C#.
No prior experience with F# is required - the tutorial will cover basic concepts of the language.
The tutorial is going to have plenty of references to the awesome [fsharpforfunandprofit.com](http://fsharpforfunandprofit.com) WebSite, which includes plenty of articles about F# written by Scott Wlaschin.

For most of the following sections there will be a direct link to a specific tagged commit that contains implementation of the application up to the point.
This allows you to follow along the process of creating the app, and get back on the track in case of any ambiguity.

Visual Studio 2013 is used throughout the tutorial, but of course you can use IDE of your choice.

You're more than welcome to create issues or pull requests for this tutorial.
The code for the application can be found
[here](https://github.com/theimowski/SuaveMusicStore), and the contents of this book are available [here](https://github.com/theimowski/SuaveMusicStoreTutorial/blob/master/en/SUMMARY.md).